trezorlib/transport: make changes to support being a separate submodule, move common functions to superclass
